refactor(template-compiler): new parser arch
