+++ glass-easel-template-compiler/src/binding_map.rs
