+++ glass-easel-template-compiler/src/lib.rs
@@ -9,13 +9,14 @@ pub use tree::TmplTree;
 pub(crate) use tree::*;
 mod expr;
 pub(crate) use expr::*;
+pub mod parse;
 mod parser;
+mod binding_map;
 mod group;
 pub use group::*;
-mod binding_map;
 mod entities;
 mod escape;
 mod js_bindings;
 #[cfg(feature = "c_bindings")]
-pub mod cbinding;
+// pub mod cbinding; // TODO
 mod proc_gen;
